exception files

+++ HotelBookingManagementSystem/src/main/java/com/Team4/project/controller/HotelController.java
@@ -2,7 +2,7 @@ package com.Team4.project.controller;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.Collectors;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -10,6 +10,7 @@ import org.springframework.web.bind.annotation.*;
 
 import com.Team4.project.entity.Hotel;
 import com.Team4.project.entity.RoomDetails;
+import com.Team4.project.exception.HotelsNotFoundException;
 import com.Team4.project.repository.IRoomDetailsRepository;
 import com.Team4.project.service.services.HotelService;
 
@@ -24,13 +25,19 @@ public class HotelController {
 	
 	@PostMapping("/")
 	public Hotel addHotel(@RequestBody Hotel hotel) {
+		Optional<Hotel> hotelById = hotelservice.showHotel(hotel.getHotel_id());
+		List<Hotel> hotelByName = hotelservice.getHotelByHotel_name(hotel.getHotel_name());
+		
+		if(hotelById.isPresent())
+			throw new HotelsNotFoundException("hotel with id "+hotel.getHotel_id()+" exists");
+		else if(!hotelByName.isEmpty())
+			throw new HotelsNotFoundException("hotel with this name already registered");
 		
 		List<RoomDetails> room = new ArrayList<>();
 		for(RoomDetails a:hotel.getRoom_details()) {
 			RoomDetails r = roomrepo.findById(a.getRoom_id()).get();
 			room.add(r);
 		}
-		
 		hotel.setRoom_details(room);
 		hotelservice.addHotel(hotel);
 		return hotel;
@@ -43,7 +50,11 @@ public class HotelController {
 	
 	@PutMapping("/{id}")
 	public ResponseEntity<Hotel> updateHotel(@PathVariable("id") int id, @RequestBody Hotel hotel){
-		Hotel updateHotel = hotelservice.showHotel(id);
+		Optional<Hotel> hotelById = hotelservice.showHotel(id);
+		if(hotelById.isEmpty())
+			throw new HotelsNotFoundException("hotel with id "+id+" not exists");
+		
+		Hotel updateHotel = hotelById.get();
 		
 		List<RoomDetails> room = new ArrayList<>();
 		for(RoomDetails a:hotel.getRoom_details()) {
@@ -67,20 +78,25 @@ public class HotelController {
 	
 	@DeleteMapping("/{id}")
 	public String deleteHotel(@PathVariable("id") int id) {
+		Optional<Hotel> hotel = hotelservice.showHotel(id);
+		if(hotel.isEmpty())
+			throw new HotelsNotFoundException("hotel with id "+id+" not exists");
 		return hotelservice.removeHotel(id);
 	}
 	
 	@GetMapping("/{id}")
 	public Hotel getHotelbyId(@PathVariable("id") int id) {
-		return hotelservice.showHotel(id);
+		Optional<Hotel> hotel = hotelservice.showHotel(id);
+		if(hotel.isEmpty())
+			throw new HotelsNotFoundException("hotel with id "+id+" not exists");
+		return hotel.get();
 	}
 	
 	@GetMapping("/name/{name}")
 	public List<Hotel> getHotelbyName(@PathVariable("name") String name) {
-		List<Hotel> hotel = hotelservice.showAllHotels();
-		
-		List<Hotel> hotelName  = hotel.stream().filter(e -> e.getHotel_name().equalsIgnoreCase(name)).collect(Collectors.toList());
-		
-		return hotelName;
+		List<Hotel> hotel = hotelservice.getHotelByHotel_name(name);
+		if(hotel.isEmpty())
+			throw new HotelsNotFoundException("Hotel does not exist");
+		return hotel;
 	}
 }
